refactor+docs: relocate tied-weight helpers, restructure dLLM CHANGELOG bullet
Move three private helpers out of unified_export_hf.py into the more
specific util modules suggested in the round-2 review:

- sync_tied_input_amax → modelopt/torch/export/quant_utils.py
  Quantizer-amax merge helper, fits the file's "Utils for quantization
  including scaling factors adjustments" charter (sibling to
  adjust_attn_amax_values, get_*_scaling_factor, etc.). Uses the
  file's existing `from warnings import warn` rather than a new
  warnings import.

- _collect_canonical_tied_patterns → modelopt/torch/export/model_utils.py
- _reorder_canonical_first       → modelopt/torch/export/model_utils.py
  Both are model-introspection helpers that walk the model graph and
  read HF's _tied_weights_keys declaration, which fits the file's
  "model type detection and classification" charter (sibling to
  get_model_type, is_multimodal_model). Added `import re` to support
  the regex compilation in _collect_canonical_tied_patterns.

unified_export_hf.py now imports the three helpers from their new
homes and uses them unchanged. No behavioral change.

tests/unit/torch/export/test_unified_export_hf.py updates its imports
to pull the three helpers from the new locations.

CHANGELOG.rst: restructure the dLLM tied-weight entry from a single
dense paragraph into a parent bullet with four sub-bullets (dedup,
sync_tied_input_amax, decoder-canonical reorder, DiffusionGemma
model-specific recipe) for readability.
